Add top level executor
The Executor coordinates running code from a source on a VM.

One of the main advantages over the previous method is that it's now
possible to display the source line on errors (at least for scan and
parse errors).

It also removes the duplication of error handling from the REPL and
Runner. Errors are now handled and reported consistently.

Other changes:

- The -> symbol is now *always* used to indicate the start of a new
  block/scope. This allows for some simplification in the Scanner.
- Formatting of $ strings is disabled for now. The plan here is to use
  a completely different strategy for $ strings (similar to Python's
  handling of f strings).
- Removed `elif` keyword. `else if` will be used instead.
- Added `export` keyword. Currently unused, but the plan is to use this
  to explicitly export objects from modules (when modules are
  implemented).