Add FizzBuzz example script/module
The purpose of this is to demonstrate and exercise a variety of features
in a somewhat realistic but bite-sized example.

Getting this to be runnable will require a number of features to be
implemented. Once the initial version is runnable, it can be made more
and more complex with the introduction of additional features such as
imports, custom types, using `match` instead of `if`, etc.