Reimplement format strings & improve tuples
When a format string is encountered in the scanning phase, it's
scanned into string and token parts, recursively. In the parsing phase,
all these parts are parsed into expressions. In the compilation phase,
these expressions are handled as other expressions and then a final
`MAKE_STRING(N)` instruction is added. When the VM runs this
instruction, the top N items are popped from the stack and joined
together into a single string.

This approach gives the format string expressions access to the runtime
context (esp. vars) so they can be evaluated dynamically.

A similar approach is now used for tuples too such that they can now
contain variables.

Other changes:

- Expressions in format strings now use `{...}` instead of `${...}`;
  this makes them easier to type and much less noisy
- Simplify `Scanner`
- Add `Stack::pop_n()`
- Add `VM::pop_n()`
- Add `MakeString` and `MakeTuple` instructions
- Simplify `VM`