Separate storage of constants from vars
This is start at memory management.

Previously, all objects were stored in a single constant storage
container and nothing was ever freed.

Now, vars are stored in the namespace for the scope they're defined in.
When a scope is exited, its namespace, and everything it contains, is
"freed" by clearing its internal storage and popping it off the
namespace stack.

Also fixed an issues with scopes not being exited when jumping due to
breaks and continues.