Implement native functions
This adds the machinery for handling native functions--that is, Rust
functions that are exposed through the runtime context.

All of the special handling for `print()` was removed as it is now
implemented as a native function and handled just like any other
function in the parsing and compilation phases.

This also moves the implementation of user-defined functions forward
a bit.