Revamp VM stack to support temp & return values
Previously, function return values were add to the constant pool,
meaning they'd live forever. Now, return values are stored directly in
the stack itself, so when a return value is popped and no longer
referenced, it should now be freed.

In addition, computed values, such as the result of binary operations,
are now stored as temporaries rather than constants so they will be
automatically freed like return values.

An enum of value stack types to allow the value stack to easily refer to
different kinds of values--constants, vars, temporaries, and return
values. When popping from the stack, the stack type is checked and the
value is pulled from the corresponding location.

- Removed unused POP instruction
- Removed PUSH instruction; this did the same thing as the LOAD_CONST
  instruction, so it was confusing to have both and the latter is more
  explicit
- Added TRUNCATE instruction
- Added VM functions for pop and peeking objects; these make retrieving
  objects a simple one-step process in most cases
- When exiting a scope, the stack is now truncated back to its size
  before entering the scope in order to clear any unused items added by
  the scope
- Use TRUNCATE in the REPL to clear the stack after eval (so it doesn't
  keep growing with inaccesible items)