Disuse Mutex for object and type refs
This seemed like a possibly good idea because it would make objects
thread safe, but it's really easy to get deadlocks when doing things
like comparing an object to itself, and it would be a *lot* of work to
get around that.

Additionally, I'm not sure making objects thread safe by default like
this is the right approach, especially since threads are far from being
implemented (if they ever are).