Simplify parsing of binary operators
Rather than converting operator tokens to strings and passing the
strings through, pass the tokens through, replacing `from_str` with
`from_token` on all the operator types.

And instead of matching on operator when creating AST nodes, try to
create each operator type in turn, leaving this responsibility solely to
the operator types and removing duplication.