Improve handling of `this`
In particular, clear `this` at the end of every statement.

Thinking that another approach would be check `b` in `a.b` to see if
it's a function (or other callable type, but none exist just yet) and
then creating some kind of bound function construct with `a` as `this`.
This way, `this` wouldn't need to be tracked and cleared in the VM, but
it would require creating additional structs.