Optimize RuntimeContext a bit as indicated by flamegraph
Remove the `depth()` method which dynamically calculated the current
namespace index.

Instead, track the namespace stack `size` and `depth` as instance
variables that are updated as scopes are entered and exited.

Also mark the `current_namespace()` as inline.

These changes do appear to make a small but noticeable difference in
performance.