Make some improvements wrt. function calls
Remove the previous `this` handling since it was wonky. Probably going
to go with the "bound method" approach. The only thing this affects
currently is `Tuple.map()`.

Change the signature of `ObjectTrait::call()` to require an object to be
returned. This is so the VM doesn't have to the check for `None` and
convert to `nil` for *every* call. Not sure if this will affect
performance much, but this is more explicit anyway and doesn't cost much
effort when writing builtin functions (noting that this doesn't apply to
user functions as they're compiled such that they'll always return an
object). It also avoids having to write `Ok(Some(...))` vs just
`Ok(...)`.

Handle builtin and user funcs in a more consistent way. In both cases,
always use their `call()` method and always start a new scope with bound
arg names before the call. And always be sure to leave the return value
on the stack before exiting the call scope.