Implement a better version of function binding
When a function is accessed via dot notation, it will be bound to the
object on the left side of the dot.

An interesting consequence of this is that a function defined in
a module (not inside a type) will have its `this` set to the module when
accessed through the module. Not sure if this is useful or not, but it
seems like it could be.

The potential problem with this is that, like JavaScript, `this` won't
be defined when a function is called directly (e.g., within the same
module without the module qualifier). A potential way around this would
be to define `this = <module>` globally for every module.

NOTE: Modules aren't implemented yet, so this is all speculative.