Store constants for each code unit with the code unit itself
Previously, constants were shared across all code units (e.g., global
scope & functions). Now, global constants such as the singleton `nil`
are store in a global constants list and code unit constants are stored
local to the unit.

To support this, the `Chunk` type was renamed and fleshed out as the
`Code` struct and some new instructions for loading global constants
were added.

This also passes `argv` directly to the compiler from the executor so
that the compiler doesn't need a dependency on the runtime context,
which allows the lifetime on `Visitor` to be removed.