Improve how function calls are handled in the VM
In particular, we need to ensure all calls, whether of builtin or user
functions, are handled consistently. This is a little tricky because
builtin functions don't use the stack and because user functions can be
called from builtin functions.

This also fixes calling of `$main` which was broken in 1abda1b05b61.