Move call to dis back after loop in VM::execute()
When `dis` is called before that allows us to see the last instruction
if there's an error, but then vars and such aren't set.

Disassembly probably needs to be completely revamped. Also thinking
about a simple disassembler mode that just directly shows the output
of the compiler without any interpretation.