Pop value from stack after every *global* statement
This keeps the stack from accumulating values unnecessarily.

Note that in non-global scopes (AKA blocks), this isn't strictly
necessary because the stack is always truncated (back to its pre-scope
state) when exiting a scope. However, it would still be good to manage
the stack in blocks to avoid the stack growing too big.