Implement a "proper" version of locals
This version uses the stack to store local vars, utilizing the scope
tree structure to track locals during compilation.

This requires very careful stack manipulation and discipline, which
could be considered a good thing.

The next iteration on this should include function parameters as locals.

This also introduces a regression in the following kind of code:

    $main = () ->                # 1
        recurs = () -> recurs()  # 2
        recurs()                 # 3

The recursive call to `recurs` on line 2 results in a name error. The
issue is that functions don't have access to the outer scope's locals so
they need to be captured somehow.  This happens because functions are
compiled separately and can't "see" the enclosing scope.