Include function args and this in locals
In order to account for jumps, this also reverts back to the stack
truncation strategy when exiting a scope rather than popping N locals.
When a jump occurs, their may be less than N locals on the stack, so we
can't rely on an exact count. The truncation strategy is easier to grok
anyway.