Improve error reporting, esp. wrt. runtime errors
- Save source lines as they're read so they can be referenced later
- Add `STATEMENT_START` instruction, which includes the statement start
  and end location
- In compiler, add `STATEMENT_START` instruction before every statement
- In VM, save statement start location when `STATEMENT_START`
  instruction is run
- When a runtime error is encountered
  - Retrieve the start & end location of the current statement
  - Show the statement line and error location

For runtime errors, this will highlight the entire statement where the
error occurred rather than pinpoint the exact location of the error, but
that's better than before where nothing was shown at all.

In order to pinpoint errors, additional location info would need to be
included somehow. This could be added for each instruction, or perhaps
a `EXPR_START` instruction could be added as well.

Since `Source` is now saving all lines anyway, it might be better to
just read the whole source up front and remove some of the complexity of
buffering lines and so forth.