Fix parsing bug that would allow statements such as `f () -> 1`
This would be parsed strangely because the `f()` would first be
interpreted as a call and simultaneously as the anonymous function
definition `() -> 1`, with the definition being passed as an arg to the
call.

At runtime, this would cause a name not found error for `f` because `f`
was never defined.