Check number of args passed to calls
This check was inadvertently removed in 9b1cf0eb6ac5. That commit
added an arg check specifically for *methods*, but the check is still
needed for other functions.

Additionally, since the VM check will always run before the check in
`make_meth!`, the latter was removed.

This also moves the arg check before everything else since there's no
need to proceed if the args don't match.