Push temporary result after in-place op
Previously, the result of an in-place op would be pushed as a var or
local type, depending on the LHS type, but this is semantically
incorrect as the result isn't a var or local.

In addition, locals on the stack are treated specially (in particular,
they're never popped), so we need to be careful not to put additional
locals on the the stack.