Make all jumps relative & remove unused jump types
Using relative jumps allows the compiler some flexibility in inserting
instructions, although care has to be taken that instructions aren't
inserted between a jump and its target (whether forward or backward).

The impetus for this is to allow instructions to be inserted at the
beginning of code segments for functions.

The following unused jump types were removed:

- JumpIf
- JumpIfElse