Initial version of closures with many limitations
Currently, only one level of enclosure works correctly. A workaround is
to pull outer vars into the local scope like so:

    f = (x) ->
        g = (y) ->
            # LHS x is a local var, RHS x is f's x parameter
            x = x
            h = () -> (x, y)

This sort of points the way to how up-linking captured vars could work.

Here's another example that doesn't work:

    $main = () ->      # 1
        f = () -> f()  # 2
        f()            # 3

`f` partilally captures itself, but it's still uninitialized (i.e.,
`nil`) at the moment `f` is called on line 3, thus leading to an error
when it's called on line 2 (since `nil` can't be called).

Part of what's making this complicated is block scoping. It's not easy
to update instructions when a captured var is found. For example:

    f = () ->            # 1
        x = 1            # 2
        block ->         # 3
            x = 2        # 4
            g = () -> x  # 1

In this example, we want to convert `x` on line 4 in the block to store
`x` into a cell and any following references to that particular `x` to
store into or load from the cell. The outer `x` on line 2 shouldn't be
touched since it hasn't been captured.

Conceptually, this isn't complicated, but in the compiler, there's
currently no easy way to update instructions only within a particular
*block* scope.