WIP: closures pt. ii
A new version of closures that threads captured vars from higher up down
to the current function. For example:

    f = (x) ->
        g = (y) ->
            h = () -> (x, y)

Previously, `x` wouldn't make its way down to `h`, but now it does.

Known issues with this version:

- When a function is called, cells are created for *all* args, even for
  those that aren't captured.
- When a function references itself, there's a little hack to keep this
  from causing an error due to the function being *declared* but not yet
  assigned at the moment of capture.

This also makes and documents the decision that vars defined *after*
a function definition won't be captured. This avoids complexit and it
seems more intuitive that such vars would *not* be captured. I'm not
sure if there's some advantage to allowing this that I might be missing.
When I think of how I've used this "feature" in other languages, it's
never seemed essential or even particularly useful (but I may be
forgetting some esoteric scenario where it actually is useful).

Additionally, this does away with storing locals on the stack. Although
that was a cool exercise, it makes things a bit more complex for no real
gain at this stagek. Performance seems good enough and not having to
deal with that extra bit of complexity will make it easier to add actual
features (e.g., this new version of closures was easier because of
this).