Make a bunch of changes aimed at fixing the REPL (WIP)
At some point recently, the REPL was broken in that globals were no
longer accessible. This is because names are now validated during
compilation whereas they were only checked at runtime before.

The essential idea here is that the REPL is an "open" module. That is, we
create a REPL module and execute everything in that context, in
particular maintaining global vars.

This initial version moves toward that but is still pretty wonky. One
thing I think needs to be done is making the VM aware of the
module it's currently executing and then global vars can be added to the
module dynamically as they're created.

Or something along those lines. This version jams the globals into the
REPL module after each eval, which doesn't seem quite right.

Additionally, these jammed-in globals still aren't then accessible in
the compiler when evaling the next bit of code.

This also makes a bit of progress toward being able to import user
modules by splitting up some of the code in `Executor` for reuse.

It also removes the wonky `keep_halt_on_top` flag from `Executor` and
the places where that trickled down.