Always pass argv through
`main()` passes it through to the various run functions, which pass it
through to the `Executor`, which passes it down to `system` module init,
where it is now added as an attribute to the `system` module.

Currently it's not possible to retrieve argv from the command line
*except* when running a file/script (including when using - as a file
name for stdin). This is because the CLI program will interpret the
first positional CLI arg as a file name and abort with an error.