Append NOOP to compiled modules
This ensures that modules end with a jump target in case they end with
a if-only conditional.

It also just generally assures that modules have an end that's not part
of any user code.

With scripts, this isn't strictly necessary since they always end with
a halt instruction, but it doesn't hurt to have it there.