Required object ref to be passed to ObjectTrait::get_attr()
This allows the "receiver" (AKA `this`) to be returned as the result
without having to clone the inner object.

Initially, this is to support the forthcoming, experimental `result.err`
pattern.

For symmetry, `set_attr()` also requires the object ref to be passed.

TODO: See if there's better way to do this that doesn't require passing
      the object ref wrapping `self` like this because it's somewhat
      redundant (and maybe indicates a bit of code/architecture smell).

NOTE: The use of the name "this" for this is kinda silly, but it matches
      how "this" is being used in method definitions.