Fix boolean evaluation of errors
When an error is evaluated directly, it's bool value is `false`:

    result = assert(false)
    !!result  # -> false

When an error is evaluated via the `.err` attribute, it's bool value is
`true`:

    result = assert(false)
    !!result.err  # -> true

One case where this could be wonky is when the result value is a bool
(or `nil`):

    # Assume this could return a bool or an error
    result = some_func()

    if !result ->
        "Is this an error or did some_func return false?"

Error checking should really only be done by checking `result.err` or
`result.ok`, but currently this can't be enforced. A possible solution
would be to disallow direct boolean evaluation of errors and only allow
it via `result.err`.