Add initial type hinting
All this does is look for something that looks for a type hint after an
identifier and then completely ignore it.

Because of the syntax this uses:

    f: Int = (s: Str) -> s.length

the existing label syntax had to be modified slightly because it's hard
to distinguish whether `label_or_name` in this form is a label or a
name:

    label_or_name: x = 1

It might be possible to do some kind of analysis to decide what to do
here, but it's much simpler to just use a slightly different label
syntax, which also has the nice effect of making labels more obvious:

    :definitely_a_label: x = 1

Another option would be to allow a label to be the only item on a line:

    label:
    x = 1

but this would be inconsistent with all the other places where inline
expressions are allowed (plus, I don't like having to do that when the
expression is simple).