Resolve parsing ambiguity by using => to start function scopes
Previously, in a form like this, the right hand side of the `||` would
be parsed as a function and end up throwing an error because `b && c`
isn't an identifier:

    if a || (b && c) ->
        "do stuff"

This was because the decision to parse a parenthesized expr as function
args was based solely on whether the expr was followed by a scope start
token `->`.

There are perhaps other ways to resolve this ambiguity, but using a
separate token is probably the simplest, as it doesn't require any
contextual knowledge. In addition, using a separate token makes function
definition stand out a bit more.

When custom types are implemented, a yet another token may need to be
used for the type def scope or a `type` keyword might be needed (or
something along these lines):

    type MyType = () =>

        @new = () =>
            "make a new MyType"

        my_method = () =.
            "do stuff"

On the other hand, it's not clear if the parens are needed in the type
def, so perhaps this could work:

    MyType =>

        @new = () =>
            "make a new MyType"