Track inline scope type (block or func) in scanner
This isn't used for anything currently, but it seems like it will be
useful.

This was added while trying to find a solution for this, which will
fail:

    match () => nil ->
        * -> nil

unless the function is parenthesized:

    match (() => nil) ->
        * -> nil

For simple cases, this could be solved by exiting the function's inline
scope when encountering the `->`, but this doesn't work for more complex
functions:

    match () => block -> nil ->
        * -> nil

The solution for this would probably require some significant reworking
of the scanner and parser, but I'm not sure it's worth the effort, esp.
since the places where this would happen[1] typically won't/shouldn't
have complex expressions like this.

[1] E.g., the condition expression for an `if`, `loop`, or `match`.