Improve main & config handling
Added a global config object to use instead of environment variables.

Environment variables are parsed through `main` and used to set global
config, which can then be accessed wherever it's needed.

This is more robust than accessing environment variables directly since
the config object validates values when they're set and retrieved.

This also improves how the REPL history path is handled. In particular,
the default is now `~/.config/feint/repl-history` and parent directories
are created if they don't exist.

Additionally, when using `-c` to run a snippet of code, `argv` is now
handled correctly instead of throwing an error.