Don't add duplicate constants to code units
When adding a constant, if it's immutable and was already added, skip
adding the constant and return the existing constant index.

What this does is discard constants that have already been created and
keep them from taking up memory.

I'm not sure how useful this really is (i.e., how much memory it saves).
When it comes to strings, they should probably be interned globally.

Also, something to note is that certain objects are never added as code
unit constants:

- Shared ints in the range [0, 256]
- Closures (these are only created in the VM)
- Bound functions (likewise, these are only created in the VM)
- Builtin functions, types, etc