Add throw keyword and instruction
The intent of this is to allow user code, in special cases, to throw and
cause the interpreter to bail out and halt.

Their are two use cases that prompted this: exiting in general and
exiting due to a failed assertion. This functionality will be used when
these functions are reimplemented in FeInt.

Nothing sophisticated happens when throwing currently. There's no stack
unwinding or anything like that. Also, *any* object can be thrown,
although it's not clear that this is particularly useful.

`throw 1` is a synonym for `system.exit(1)`.