First pass at loading modules up front rather than at runtime
This is somewhat complex, so there are likely to be issues.

One known issue is with the way module globals are loaded from the
runtime context, which can include globals from other modules. The fix
for this is to run in a module context or have a stack of contexts or
something along those lines.

Overall, this is a major simplification of module loading since
everything is now coordinated in the `Executor` instead of bouncing
around. This also means the wonky global config object is no longer
needed, since the `Executor` holds the config.