Always reset VM before executing a module
Note that a module is *always* the top level unit of execution.
Imported modules are loaded *up front* during the initial setup phase,
not at runtime. At runtime, there will never be a switch to executing
another module--there will only be switches to running other functions
(including methods).

One issue relating to this is what is considered the global context
within a function. Currently, this will always resolve to the top level
module that is the unit of execution described above. This is incorrect
in that it's typically expected that the global context within
a given function is the module where the function is defined.