Simplify builtin module discovery & installation
Include builtin modules in the binary so that there essentially is no
discovery process.

This also simplifies the installation process in that no decision needs
to be made about where on the file system to copy the module files and
there's no need for a convoluted discovery process that differs between
debug, test, and release.