Replace system.modules w/ MODULES global during bootstrap
Previously, `system.modules` and `MODULES` were two separate `Map`s that
needed to be kept in sync, but when adding a module dynamically from
FeInt code like this:

    import std.system
    module = Module.new("name", "<path>", "doc", {"name": "val"})
    system.add("name", module)

the new module wouldn't be added to `MODULES`, meaning that it would be
inaccessible to the interpreter.

Now, `system.modules` (the one defined in `system.fi`) is replace with
`MODULES` during bootstrap, so there is no issue with syncing.

This also allows for some redundant code to be removed from `Executor`.