Replace "builtin" with "intrinsic" where that makes sense
The distinction here is that "intrinsic" means implemented in the host
language, Rust, while "builtin" means a module, type, function, etc that
is provided by the system and that may be implemented in Rust or FeInt
or a combination.

For example, the `std` module is partly implemented in Rust and partly
implemented in FeInt. The base Rust module is "intrinsic" but the module
as a whole, including functions implemented in FeInt, is a "builtin".