First pass at reorganizing into workspace w/ multiple packages
The impetus for this was wanting to add a package with `derive` macros,
since those need to be in their own package, and that got me thinking
that it might be good to break things up into logical packages and also
learn about workspaces generally.

The hardest part of this was breaking circular dependencies, and there's
still one issue related to that that is unresolved...

Most things work, but intrinsic functions that were previously passed
a reference to the VM so they could make calls are broken because the VM
reference is no longer passed into such functions.

The VM is no longer passed in because 1. types shouldn't be aware of the
VM and 2. the `feint-vm` package uses the `feint-builtins` in package,
so `feint-builtins` can't also have `feint-vm` as a dependency.

One solution that comes to mind is to return some kind of lazy object
from intrinsic functions that the VM would then evaluate. This would
only apply to the return values of intrinsic functions, since this is
the only place where this is an issue.

Something else that's wonky is how types indicate errors when an
operation isn't defined. Previously, they would directly throw
a `RuntimeErr`, which would cause the VM to halt (which was arguably
incorrect anyway).

Now, they all return `None` when an operation isn't defined on a type or
between different types, and the associated error checking is handled in
the VM. It might be better to instead make all the operations always
return an object, which could be an `ErrObj` instance so the VM doesn't
have to handle that (because it's somewhat awkward to handle them
there).

Type checking during compilation would be another way to fix this issue,
but that would be a massive undertaking.

Lastly, some things were moved to dubious locations in order to break
dependency cycles. In particular, `Code` and `Inst` were moved to
`feint-builtins`, in the `types` module, which doesn't seem like the
right place for them.  The issue here is that the `Func` and `Module`
types need access to `Code`, but it can't live in `feint-vm` to avoid
a circular dependency.

Some related TODOs:

- Remove `feint-vm` dep from `feint-cli` and make the driver translate
  `VMState`s to a driver result
- Remove `feint-builtins` dep from `feint-cli` and add a `create_module`
  method to the driver
- Add baked-in constants back to VM (singleton `nil`, `true`, etc)
- Move `Code` to `feint-util`?
- Rename `feint-util` to `feint-common`?