Fix less than & greater than comparisons between floats and ints
In the case of `f < i` and `f > i` (LHS float, RHS int), there was a bug
where the actual operation performed was `i < f` and `i > f`.

I'm guessing this was due to an errant copypasta (along with a lack of
test coverage).

Added new utility functions for these cases and updated the names of
the existing functions and their args for clarity.

NOTE: This is a manual cherry pick of f624692ab1a0 from dev.