When jumping, only exit scopes if jump actually happens
This applies to the various `JUMP_IF` instructions, which would always
exit scopes even if the jump didn't happen.

The reason this hasn't caused any problems is because currently all
such are always compiled such that they don't exit any scopes.

Thinking on this a bit, I'm not sure any of the jump instructions other
than `JUMP` actually need to exit scopes, since it's only necessary to
exit scopes when jumping *out* of a block and `JUMP` is the only
instruction that does that currently.