Implement reassignment via <-
`<-` is used for reassigning a var defined in an OUTER scope, excluding
globals.

`=` always creates a new var in the current scope, similar to using
`let` in Rust.

Within the SAME scope, `=` can be used for reassignment, but since there
are no `let` or `mut` keywords in FeInt, there's no way to indicate in
an INNER scope that an assignment should reassign a var defined in an
OUTER scope.

Therefore, a new operator is necessary for this purpose, which is
especially usefully in certain loops--e.g., when a flag needs to be set
or an accumulator needs to be updated.

In terms of performance, using reassignment in the Mandelbrot example
makes a noticeable difference of about 150ms per run using the following
command on an M1 Mac:

    time rr examples/mandlebrot.fi 50 50

Using reassignment, this runs in ~500ms. The previous version, which
used in-place update hacks, took about ~650ms.